_backend/requirements.txt
asgiref==3.7.2
boto3==1.28.54
botocore==1.31.54
certifi==2023.7.22
charset-normalizer==3.2.0
coverage==7.3.2
Django==4.2.5
django-appconf==1.0.5
django-cors-headers==4.3.0
django-dotenv==1.4.2
django-storages==1.14
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
idna==3.4
jmespath==1.0.1
packaging==23.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
rcssmin==1.1.1
requests==2.31.0
rjsmin==1.2.1
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.16
